Additional code cleanup
