--- Bausch Take Home/Models/Categories.swift
@@ -1,12 +1,0 @@
-//
-//  Categories.swift
-//  Bausch Take Home
-//
-//  Created by Tim Bausch on 1/10/22.
-//
-
-import Foundation
-
-struct Categories: Decodable {
-    var categories: [Category]
-}
