--- Bausch Take Home/Models/MealDetail.swift
@@ -1,12 +1,0 @@
-//
-//  MealDetail.swift
-//  Bausch Take Home
-//
-//  Created by Tim Bausch on 1/10/22.
-//
-
-import Foundation
-
-struct MealDetail: Decodable {
-    var meals: [Meal]
-}
