--- Bausch Take Home/Models/Meals.swift
@@ -1,12 +1,0 @@
-//
-//  Meals.swift
-//  Bausch Take Home
-//
-//  Created by Tim Bausch on 1/10/22.
-//
-
-import Foundation
-
-struct Meals: Decodable {
-    var meals: [MealsInCategory]
-}
