--- Bausch Take Home/Models/MealsInCategory.swift
@@ -1,14 +1,0 @@
-//
-//  MealsInCategory.swift
-//  Bausch Take Home
-//
-//  Created by Tim Bausch on 1/10/22.
-//
-
-import Foundation
-
-struct MealsInCategory: Decodable {
-    var strMeal: String
-    var strMealThumb: String
-    var idMeal: String
-}
